pyproject.toml
[project]
name = "wartools-net"
version = "0.1.0"
description = "Build scripts for WarTools.net"
requires-python = ">=3.14"
dependencies = [
    "commonmark>=0.9.1",
    "pillow>=10.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.4",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
